pyproject.toml
[tool.poetry]
name = "workshop-dw"
version = "0.1.0"
description = ""
authors = ["Bruno Silva <bruno.ogs@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
yfinance = "^0.2.51"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
